Fix training

+++ arenai_agent/tests/include/arenai_agent_tests/tests_distributions/tests_truncated_normal.h
@@ -7,8 +7,8 @@
 
 #include <gtest/gtest.h>
 
-typedef int UpperBound;
-typedef int LowerBound;
+typedef float UpperBound;
+typedef float LowerBound;
 
 typedef std::vector<int64_t> Shape;
